edit i delete deluju ok za sada; jos uvek nisu sigurni

+++ src/main.js
@@ -21,7 +21,7 @@ app.on('ready', () => {
     win.webContents.openDevTools();
 
     ipcMain.on("list-Artikl", () => {
-        database.db.all(`SELECT SifraArtikla, Naziv, JedinicaMere, Cena
+        database.db.all(`SELECT ID_Artikla, SifraArtikla, Naziv, JedinicaMere, Cena
             FROM Artikl`, (err, rows) => {
                 if (err) {
                     throw err;
@@ -47,22 +47,42 @@ app.on('ready', () => {
     });
 
     ipcMain.on("insert-Artikl", (evt, SifraArtikla, Naziv, JedinicaMere, Cena) => {
-        database.db.run(`INSERT INTO
-                        Artikl (SifraArtikla, Naziv, JedinicaMere, Cena)
-                        VALUES (?, ?, ?, ?)`, [SifraArtikla, Naziv, JedinicaMere, Cena], (err) => {
+        database.db.run(`INSERT INTO Artikl(SifraArtikla, Naziv, JedinicaMere, Cena)
+                        VALUES(?, ?, ?, ?)`, [SifraArtikla, Naziv, JedinicaMere, Cena], (err) => {
                             if (err) {
                                 dialog.showErrorBox('Greska pri unosu podataka', err.message);
+                            } else {
+                                win.reload();
                             }
-                            win.reload();
                         });
     });
 
-    ipcMain.on("delete-Artikl", (evt, SifraArtikla) => {
-        database.db.run(`DELETE FROM Artikl
-                        WHERE SifraArtikla=?`, SifraArtikla, (err) => {
+    ipcMain.on("delete-Artikl", (evt, ID_Artikla) => {
+        let options = {
+            buttons: ["Da", "Ne"],
+            message: "Da ste sigurni da zelite da obrisete?"
+        };
+        let response = dialog.showMessageBoxSync(options);
+        if (response == 0) {
+            database.db.run(`DELETE FROM Artikl
+                            WHERE ID_Artikla=?`, ID_Artikla, (err) => {
+                                if (err) {
+                                    dialog.showErrorBox("Greska pri brisanju", err.message);
+                                } else {
+                                    win.webContents.send("deletedRow");
+                                }
+                            });
+        }
+    });
+
+    ipcMain.on("edit-Artikl", (evt, ID_Artikla, SifraArtikla, Naziv, JedinicaMere, Cena) => {
+        database.db.run(`UPDATE Artikl
+                        SET SifraArtikla=?, Naziv=?, JedinicaMere=?, Cena=?
+                        WHERE ID_Artikla=?`, [SifraArtikla, Naziv, JedinicaMere, Cena, ID_Artikla], (err) => {
                             if (err) {
-                                dialog.showErrorBox("Greska pri brisanju", err.message);
+                                dialog.showErrorBox("Pogresno uneti podaci", err.message);
                             }
+                            win.reload();
                         });
     });
 });
